requirements.txt
flask==2.0.1
requests==2.26.0
flask-bootstrap5==0.1.dev1
psycopg2-binary==2.9.1